BF: Wrong postprocessing order in clone
Before we can deal with special remotes, we need to figure it's an annex
and initialize it.
Also: Local config settings in postprocessing don't make sense, if we
only afterwards double-check that we do indeed have an installed dataset
as a result, etc.
Furthermore, there's a possible bug in annex, where git-annex-info
(indirectly called by postclonecfg_ria) half-initializes the repo
preventing us from later detect, that we still need to do it
(particularly in order to autoenable special remotes).
See issue #4152.