No creation by special remote
Special remote now expects an existing setup at the remote end.
It was a bad idea to begin with, since initremote are even more so
prepare are called in all kinds of situations leading to
unintended creation (trials). Furthermore we need autoenable in a
clone to fail on non-existing/non-accessible remote end.