RF: Keep read config separated by source
This change is motivated by gh-4813, where a dedicated ConfigManager
instance needs to be built only to be able to determine if the local
repo config has a particular setting (while ignoring the committed
config). With the changes here it becomes easier to enhance the class
with source-aware query methods.

For the time being, a merged config representation (identical to the
only representation there was before) is kept as a dedicted dict. This
is done in order to avoid runtime cost for multi-dict lookups.

The internal book-keeping  in the class has also been changed to ease
RF and save runtime. Based on that mtime tests are now ran tailor
to each potential git-config call, only triggering it when the relevant
files did change. This is in contrast to the previous behavior that
always triggered two git-config calls whenever any file changed.