BF: Avoid ConfigManager instance creation on each command execution
This was done to have a safe way of looking for result hooks. This
change keeps thiings safe, but avoids loading config from scratch
just for this single query. Instead, it used the separate internal
representation of local and committed config to perform a constrained
query. This is almost as fast as a normal query in all standard cases
(standard being: very few keys in the committed config).

Fixes gh-4813